Trying a slower clock (2MHz)
This will allow a PWM freq of 976, doubled from previous, and an
effective PWM freq of 488 - high enough to avoid flickering and low
enough to play nice with the MOSFETs. Hopefully. This is something
that'll just have to be tested on the final board.

Also reassigned bank switching duty to the TIM0 interrupt.